* this fixes a slew of warnings and errors we were getting on rinkeby, also it fixes making repeat calls to coingecko when we already know that they don't have an image for the contract address we're querying * cleanup large constants
