+++ lib/components/IndexUI.jsx
@@ -3,7 +3,8 @@ import Link from 'next/link'
 import { motion } from 'framer-motion'
 import { useRouter } from 'next/router'
 
-import { ANIM_BANNER_VARIANTS, POOL_LIST_TABS } from 'lib/constants'
+import { ANIM_BANNER_VARIANTS } from 'lib/constants/framerAnimations'
+import { POOL_LIST_TABS } from 'lib/constants'
 import { useTranslation } from 'lib/../i18n'
 import { useReducedMotion } from 'lib/hooks/useReducedMotion'
 import { usePools } from 'lib/hooks/usePools'
