+++ lib/constants/communityPoolsBlockList.js
@@ -1,0 +1,126 @@
+export const COMMUNITY_POOLS_BLOCK_LIST = {
+  1: [
+    '0xfb99ac6fdd87bc749dc32d772624c06c50b28a34',
+    '0xcc75e748342fa103099cf32c09ebda39b4556f61',
+    '0x9bdd57c278794bdced35f091bb26736a4cf4aaa6',
+    '0x8932f3e02bdd4caa61bdc7be0a80dd2911a78071',
+    '0x22964f820d40f22f59bf4a7f06aa8f27b159e398',
+    '0x908d0f3bcdb4ece202f49bfd5037e1bdb4d57fd2',
+    '0x0725ca260292d3c3911d7b67d525682ec861339b',
+    '0x1ee6194b0e28c12df0fd107e5eb883aa071f8c9a',
+    '0x258d1ab585593831f2dcc898722fb924fc0e3609',
+    '0x3dd1749e00b6376779c5e63a38f7b5bbbd7ce3f2',
+    '0x440a0622f124e4fd9925dac51b3bf1907e95f55b',
+    '0x4b33763af4886278bfdf870dd7ea4838620bc7e9',
+    '0x53de5a70527c0117c81319db4a05967f469a32f0',
+    '0x4b912b18a94a1640c4230f9444c3871a6f4f9d5e',
+    '0x7339b77606b96bd743f7c4646e51b024e47d7eff',
+    '0x8f961526bc2fb2a55287522e728a7717d78ecc19',
+    '0xb2153afa4b1bf7bfd89a3c94df5acc13d93558f7',
+    '0xcc6f8a4b3f04c688ea8681eb5ba33a5bc2f9df21',
+    '0x3e2e88f6eaa189e397bf87153e085a757028c069',
+    '0x91d7bbd2a8dfe4be5b88d4224b461de9b36d2aa9',
+    '0x55ab6e07fb3ae2598fe463ea4dcf872a006fec77',
+    '0x3bdbdb4b14f599dcd7d79ea48aff1f064fecc31e'
+  ],
+  4: [
+    '0x252396a74e71538a96e3e92f1a9d705b164b2f5b',
+    '0x32798f0da70bcc706cae8417d7c64140fbd53100',
+    '0x6b9fff203dc9db99451a2e27ace9e713355fdf75',
+    '0x00c5d25a5002dad8e5dcc31eef7e54a602beb260',
+    '0x013f9e02cddedf88f5590b4bba3884e16030fce0',
+    '0x033f62f10beff208e0fbb789a97c41dc7765bf78',
+    '0x038a7b05bae5ec0006b35924d17c96d625c03672',
+    '0x0c8b1d618331e9eccfaa8a9c6481408eaa54e68c',
+    '0x127a2523a1494cc17748f431a5b110d9c7302042',
+    '0x12e8fb4d5b4c5a8d2a99c78f889b08ae11a4aab5',
+    '0x14ffe268a0ef2c94926670b808aef0bf55f7450f',
+    '0x16783e9fe1718f0423893ba71157de7f409d4d52',
+    '0x16b9bda75bb19a949dbaceb4214c37c178116d2e',
+    '0x1704d913abf7ca5b9f4ad4a3e2c39eca1cc966e0',
+    '0x190656c3f12b6c75b453812e56892e9352eee730',
+    '0x199e9e899c09da063b3638f475b4e7cae1b4f8e0',
+    '0x19eb7053e84ff0b71b165fa23a5185362d687bcb',
+    '0x1bd6c495eb25130787b5932bae25250039044e43',
+    '0x1beeabb5dd774fdc0e38d6c9161f1ecb6f38a8dc',
+    '0x1c55490ae814c1ace086030704f5d82529a273ae',
+    '0x1df85e3bf0e68ed7fcd02d493d84e985a7b9121b',
+    '0x1efcd798f15f74ee70268c72d71999a91f45f8d3',
+    '0x1f8e600f679440a2e5d1bcc5fbad623f70f8b9f9',
+    '0x21d72fb699408b5b0606f39bb6a48a8b65c3bfe1',
+    '0x2457a1b478b6adbe860e92b5d8800be29affb1cf',
+    '0x27b88dedeca006a409198a9ed5140f3dc259aa52',
+    '0x283fd9fd9e13cb16606b1035e83c4ca1d24c2818',
+    '0x2b2430d92f06c5350ff93a477b0e226c4745ab30',
+    '0x2c3f232215baaabee82a661cbc481b2fb2e7779f',
+    '0x302e9ad0746c04f3c34e334bc5dd870a6c7f7e04',
+    '0x3700f26a48e3a65765981211ad33ea5fd2d4695b',
+    '0x370a54627a471148311dce51afb3d51851fedc54',
+    '0x37ba7acab4074bfad7bd2d7706b7fc7339f75f37',
+    '0x3b997dc80892da50ba16382a245210d141f116fb',
+    '0x3bc9f20989b28847b6bd7cd644936e747c4eaca3',
+    '0x401a279c88a3e394d83b7fb589b4c37725e8db65',
+    '0x41812c7c7fe363c2a8908ac5d7bb8f7c2f9de834',
+    '0x41ee149372238fbce0f3c5e7076aa253d0fc4c70',
+    '0x433ebbd967a81e49f98b9c0acd8946183c9b851f',
+    '0x4b24865ad142721e68999dab5eb6f12439d2c1c0',
+    '0x4d5c97433045f033c87872c7740d7488ed8eab0f',
+    '0x4dcdf0b9bf032b201e3705af6fab0df7f8769861',
+    '0x4f16ded8c022909eb7223e6f53d2c3dce07263c1',
+    '0x50fff7d3c226c412a252d4f9b64165b37347e949',
+    '0x51a2bfd7e9c2a25042029a8df01c187518c6bd4e',
+    '0x5299ff5704a6f701b650b24e8c3af9bab23e4f2d',
+    '0x56d3b9ca68e8ea9545f78fc706d1d64ac833849b',
+    '0x584d61506631c1714ba2187b3f009175462e500a',
+    '0x5ec62227a749682a02d87177692f54cb821f675a',
+    '0x5ee6a9140105b6e74368325a0cf3f1e480ca9535',
+    '0x60964fe1342b33f6c4ac6221c8966c72e27d2984',
+    '0x61b9ece8282d6a763171f8c67c9835ac7a262bfc',
+    '0x63a6c22a3d9c0e7e79b9ad99e0d8d014c224481f',
+    '0x666f4bb0a85013e2a5dfc86a1aa0fd3a98dcad17',
+    '0x673b2da80f345ec9fd4524e9fb5b342a36fd21f7',
+    '0x68a6adc8f852211290e8ec1af8f39be9459d1161',
+    '0x695758c492394a33d5aef2df0ba22158d3c2a3f5',
+    '0x6b86c780a26445209cbd19a6d59a8a11d247d0e1',
+    '0x6cc0dcffeffdfc611394f452b82407f3d0a51e18',
+    '0x71f1756c2ab6ce4df0e862ad93e2994ee387c9db',
+    '0x72758a3bf857bb7968586ad5b5a50d5e6745f24b',
+    '0x73216f8173959b6abb6e78e80bd251071787042c',
+    '0x735534aa51e0e233bd9b5a1ec371736e33af37ea',
+    '0x73c6e9976e3e2002be747291840af5ecc24c88dc',
+    '0x76a38812c733ca1341ad33ba9c085701c5df4341',
+    '0x78028d68be9d1573e7fec77045832c0372cf7876',
+    '0x78066ea14ca2219a00ded36fa30f352a59fdbb03',
+    '0x78c1554336055217adceef041c912f26d2171ef4',
+    '0x7e83065cdd7d184b02cf9a63582252aed1a4470e',
+    '0x8075f98eb2589ce69cc9a70da6e34345c2723df5',
+    '0x840cc10905b749c7a50e955c6b8a8e2593b7a5bd',
+    '0x8483d86c45db48703e16faaad5ecbffd1590a277',
+    '0x860cfafc509cd02a3196ce108f92499d28d6b52d',
+    '0x86a131d91f15f2f29d7aa9a014d33b012893aeaf',
+    '0x8c0d17691b3958e8d6ff6d44d9522110b859ba72',
+    '0x8d17f076eec480a5ac009d04cb1f33f5d5014b81',
+    '0x9362f44c3386ecba12e5c6a3885dd58f5a6c6260',
+    '0x9623f002a7a4d0c6317a3116a6ea46d29cd3708d',
+    '0x96e22522fc2da00c74227f77d976f9b878ee5a62',
+    '0x96f90c0d6d710aac895c54ff69a7f22068858114',
+    '0x96fb45276e7080f8fa62da0ec6ea255d2703816c',
+    '0x976fcc950c84446efe583e794808d364bdb4eda3',
+    '0x97a66fbd390ed9978e13b196d8c79e6e256e9a3b',
+    '0x9b72b80d14d70b99ecba97532fd39db4ee57d1e6',
+    '0x9bb6f9958ff220e7ce06c477f6a2767a4f847def',
+    '0x9e8a6669117bd693d31f8d0e64485a48b5794a85',
+    '0x9f6df0487958cf4116402f5ce6be2536d5c6ca25',
+    '0xa384e6195efce56bcad51ac3a79e87eafdfdb4d6',
+    '0xa3ba03a670fb557ce9de737096234ad3cdf03373',
+    '0xa4db12c9ed2cda277e549a3b0a16e424c91c538a',
+    '0xa5d56ed859c3ffcc64db11f9c9e58bcd4857220d',
+    '0xa9f6f79afdca69adbd1f543891ed831403af7be4',
+    '0xaffe5a0cc1f6db9b97ea1c8127835dbe77f9484f',
+    '0xb53ce598b044a93bcf1ce8f36e8ff7bbfb7aea85',
+    '0xb60931253325371853d37f84cf1559b9ff8e767c',
+    '0xb727d20b5e0413cf4e773fd5cdc5514963e53491',
+    '0xb766819565a5bb463383ae3764aeb9e2e8a54c64',
+    '0xba3f649be56408333c6ed7e159750b70fc489bea'
+  ]
+}
