+++ lib/constants/queryKeys.js
@@ -1,0 +1,34 @@
+export const QUERY_KEYS = {
+  communityPoolsQuery: 'communityPoolsQuery',
+  ethereumErc20sQuery: 'ethereumErc20sQuery',
+  ethereumErc721sQuery: 'ethereumErc721sQuery',
+  ethereumLootBoxQuery: 'ethereumLootBoxQuery',
+  ethereumPoolQuery: 'ethereumPoolQuery',
+  ethereumUsersDripQuery: 'ethereumUsersDripQuery',
+  ethereumUsersV2Query: 'ethereumUsersV2Query',
+  ethereumUsersChainQuery: 'ethereumUsersChainQuery',
+  lootBoxQuery: 'lootBoxQuery',
+  poolAddressesQuery: 'poolAddressesQuery',
+  poolQuery: 'poolQuery',
+  poolsQuery: 'poolsQuery',
+  poolDripsQuery: 'poolDripsQuery',
+  poolPrizesQuery: 'poolPrizesQuery',
+  playerQuery: 'playerQuery',
+  playerPrizesQuery: 'playerPrizesQuery',
+  prizeQuery: 'prizeQuery',
+  sponsorQuery: 'sponsorQuery',
+  uniswapTokensQuery: 'uniswapTokensQuery',
+  accountQuery: 'accountQuery',
+  prizePoolAccountQuery: 'prizePoolAccountQuery',
+  controlledTokenBalancesQuery: 'controlledTokenBalancesQuery',
+  claimablePoolQuery: 'claimablePoolQuery',
+  claimablePoolTotal: 'claimablePoolTotal',
+  retroactivePoolClaimDataQuery: 'retroactivePoolClaimDataQuery',
+  poolTokenFaucetsQuery: 'poolTokenFaucetsQuery',
+  poolTokenDataQuery: 'poolTokenDataQuery',
+  coingeckoImagesByContractAddressesQuery: 'coingeckoImagesByContractAddressesQuery',
+  coingeckoImageQuery: 'coingeckoImageQuery',
+  claimablePoolFromTokenFaucets: 'claimablePoolFromTokenFaucets',
+  claimablePoolTokenFaucetAddresses: 'claimablePoolTokenFaucetAddresses',
+  tokenFaucetDripRateQuery: 'tokenFaucetDripRateQuery'
+}
