+++ lib/constants/tokenImages.jsx
@@ -1,0 +1,65 @@
+export const TOKEN_IMAGES = {
+  '0x9d942bd31169ed25a1ca78c776dab92de104e50e':
+    '/tokens/0x9d942bd31169ed25a1ca78c776dab92de104e50e.png',
+  '0xb0dfd28d3cf7a5897c694904ace292539242f858': '/tokens/lotto.png',
+  '0x0391d2021f89dc339f60fff84546ea23e337750f': '/tokens/bond.png',
+  '0x9f7905c7bd5ec9e870ed50f0e286f2742c19f5b3': '/tokens/dpi.png',
+  '0x03ab458634910aad20ef5f1c8ee96f1d6ac54919': '/tokens/rai.png',
+  '0x6b175474e89094c44da98b954eedeac495271d0f': '/tokens/dai-new-transparent.png',
+  '0x57bc752ec42238bb60a6e65b0de82ef44013225d': '/tokens/arto.png',
+  '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48':
+    'https://assets.coingecko.com/coins/images/6319/small/USD_Coin_icon.png?1547042389',
+  '0x2260fac5e5542a773aa44fbcfedf7c193bc2c599':
+    'https://assets.coingecko.com/coins/images/7598/small/wrapped_bitcoin_wbtc.png?1548822744',
+  '0x06f65b8cfcb13a9fe37d836fe9708da38ecb29b2':
+    'https://assets.coingecko.com/coins/images/11521/small/FAME.png?1590622461',
+  '0x1494ca1f11d487c2bbe4543e90080aeba4ba3c2b':
+    'https://assets.coingecko.com/coins/images/12465/large/defi_pulse_index_set.png',
+  '0x1f9840a85d5af5bf1d1762f925bdaddc4201f984':
+    'https://assets.coingecko.com/coins/images/12504/small/uniswap-uni.png?1600306604',
+  '0x2b591e99afe9f32eaa6214f7b7629768c40eeb39':
+    'https://assets.coingecko.com/coins/images/10103/small/HEX-logo.png?1575942673',
+  // '0x2e703d658f8dd21709a7b458967ab4081f8d3d05': '',
+  '0x429881672b9ae42b8eba0e26cd9c73711b891ca5':
+    'https://assets.coingecko.com/coins/images/12435/small/pickle_finance_logo.jpg?1599817746',
+  '0x5dbcf33d8c2e976c6b560249878e6f1491bca25c':
+    'https://assets.coingecko.com/coins/images/12210/small/yUSD.png?1600166557',
+  '0x6e36556b3ee5aa28def2a8ec3dae30ec2b208739':
+    'https://assets.coingecko.com/coins/images/12380/small/build.PNG?1599463828',
+  '0x7865af71cf0b288b4e7f654f4f7851eb46a2b7f8':
+    'https://assets.coingecko.com/coins/images/7383/small/2x9veCp.png?1598409975',
+  '0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9':
+    'https://assets.coingecko.com/coins/images/12645/small/AAVE.png?1601374110',
+  '0x8ba6dcc667d3ff64c1a2123ce72ff5f0199e5315':
+    'https://assets.coingecko.com/coins/images/10972/small/ALEX.png?1586742545',
+  '0xa0246c9032bc3a600820415ae600c6388619a14d':
+    'https://assets.coingecko.com/coins/images/12304/small/Harvest.png?1599007988',
+  '0xc00e94cb662c3520282e6f5717214004a7f26888':
+    'https://assets.coingecko.com/coins/images/10775/small/COMP.png?1592625425',
+  '0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f':
+    'https://assets.coingecko.com/coins/images/3406/small/SNX.png?1598631139',
+  '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2':
+    'https://assets.coingecko.com/coins/images/2518/small/weth.png?1547036627',
+  '0xd533a949740bb3306d119cc777fa900ba034cd52':
+    'https://assets.coingecko.com/coins/images/12124/small/Curve.png?1597369484',
+  '0xe2f2a5c287993345a840db3b0845fbc70f5935a5':
+    'https://assets.coingecko.com/coins/images/11576/small/mStable_USD.png?1595591803',
+  '0x117c2aca45d87958ba054cb85af0fd57be00d624':
+    '/tokens/0x117c2aca45d87958ba054cb85af0fd57be00d624.png',
+  // '0x2e703d658f8dd21709a7b458967ab4081f8d3d05': '',
+  '0x8b9c35c79af5319c70dd9a3e3850f368822ed64e':
+    '/tokens/0x8b9c35c79af5319c70dd9a3e3850f368822ed64e.png',
+  '0xdac17f958d2ee523a2206206994597c13d831ec7':
+    'https://assets.coingecko.com/coins/images/325/small/Tether-logo.png?1602230054',
+  '0xa91ac63d040deb1b7a5e4d4134ad23eb0ba07e14':
+    'https://assets.coingecko.com/coins/images/12478/small/Bella.png?1602230054',
+  '0x08d32b0da63e2c3bcf8019c9c5d849d7a9d791e6':
+    'https://assets.coingecko.com/coins/images/850/small/dentacoin.png?1547034647',
+  '0x6b3595068778dd592e39a122f4f5a5cf09c90fe2':
+    'https://assets.coingecko.com/coins/images/12271/small/sushi.jpg?1598623048',
+  '0x0954906da0bf32d5479e25f46056d22f08464cab':
+    '/tokens/0x0954906da0bf32d5479e25f46056d22f08464cab.png',
+  '0xd291e7a03283640fdc51b121ac401383a46cc623':
+    'https://assets.coingecko.com/coins/images/12900/small/rgt_logo.png?1603340632',
+  '0x334cbb5858417aee161b53ee0d5349ccf54514cf': '/favicon.png'
+}
