+++ lib/hooks/useClaimablePoolFromTokenFaucets.js
@@ -1,14 +1,16 @@
-import { useClaimablePoolTokenFaucetAddresses } from 'lib/hooks/useClaimablePoolTokenFaucetAddresses'
-import TokenFaucetABI from '@pooltogether/pooltogether-contracts/abis/TokenFaucet'
-import ERC20Abi from 'abis/ERC20Abi'
-import { useQuery } from 'react-query'
-import { DEFAULT_TOKEN_PRECISION, QUERY_KEYS } from 'lib/constants'
-import { testAddress } from 'lib/utils/testAddress'
 import { useContext } from 'react'
-import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
-import { batch, contract } from '@pooltogether/etherplex'
 import { ethers } from 'ethers'
+import { useQuery } from 'react-query'
+import { batch, contract } from '@pooltogether/etherplex'
+import TokenFaucetABI from '@pooltogether/pooltogether-contracts/abis/TokenFaucet'
+
+import ERC20Abi from 'abis/ERC20Abi'
+import { DEFAULT_TOKEN_PRECISION } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
+import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
+import { useClaimablePoolTokenFaucetAddresses } from 'lib/hooks/useClaimablePoolTokenFaucetAddresses'
 import { useReadProvider } from 'lib/hooks/useReadProvider'
+import { testAddress } from 'lib/utils/testAddress'
 
 // TODO: Update this to read from The Graph
 // It'd be nice to copy the style of Sybil and have a large hook fetching as much as possible
@@ -49,6 +51,14 @@ export const useClaimablePoolFromTokenFaucets = () => {
 async function getClaimablePoolFromTokenFaucets(provider, usersAddress, tokenFaucetAddresses) {
   const claimableAmounts = {
     total: 0
+  }
+
+  // This rinkeby contract is bad and causes annoying etherplex/ethers errors
+  // when we try to query data against it:
+  if (provider?.network?.chainId === 4) {
+    tokenFaucetAddresses = tokenFaucetAddresses.filter(
+      (address) => address.toLowerCase() !== '0xe70c3669371be02cde0809df6d9c514804cad2f0'
+    )
   }
 
   for (const tokenFaucetAddress of tokenFaucetAddresses) {
