+++ lib/hooks/useClaimablePoolTokenFaucetAddresses.js
@@ -2,7 +2,7 @@ import { batch, contract } from '@pooltogether/etherplex'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import PrizePoolAbi from '@pooltogether/pooltogether-contracts/abis/PrizePool'
 import PrizeStrategyAbi from '@pooltogether/pooltogether-contracts/abis/PeriodicPrizeStrategy'
-import { QUERY_KEYS } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { useContext } from 'react'
 import { useQuery } from 'react-query'
 import { ethers } from 'ethers'
