+++ lib/hooks/useCoingeckoImageQuery.js
@@ -1,6 +1,6 @@
 import { useQuery } from 'react-query'
 
-import { QUERY_KEYS } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { axiosInstance } from 'lib/axiosInstance'
 
 const COINGECKO_IMAGE_BY_CONTRACT_ADDRESS_LAMBDA_PATH = `/.netlify/functions/coingeckoImagesByContractAddress`
@@ -23,6 +23,8 @@ const _getImage = async (address) => {
 export function useCoingeckoImageQuery(address) {
   return useQuery([QUERY_KEYS.coingeckoImageQuery, address], async () => _getImage(address), {
     enabled: address,
-    refetchInterval: false
+    refetchInterval: false,
+    refetchOnWindowFocus: false,
+    retry: false
   })
 }
