+++ lib/hooks/useEthereumLootBoxQuery.js
@@ -3,7 +3,8 @@ import { useQuery } from 'react-query'
 import { ethers } from 'ethers'
 import { isEmpty } from 'lodash'
 
-import { MAINNET_POLLING_INTERVAL, QUERY_KEYS } from 'lib/constants'
+import { MAINNET_POLLING_INTERVAL } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { useContractAddresses } from 'lib/hooks/useContractAddresses'
 import { useReadProvider } from 'lib/hooks/useReadProvider'
