+++ lib/hooks/usePlayerPrizesQuery.js
@@ -1,7 +1,8 @@
 import { useContext } from 'react'
 import { useQuery } from 'react-query'
 
-import { MAINNET_POLLING_INTERVAL, QUERY_KEYS } from 'lib/constants'
+import { MAINNET_POLLING_INTERVAL } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { getPlayerPrizesData } from 'lib/fetchers/getPlayerPrizesData'
 import { usePoolTogetherSubgraphClient } from 'lib/hooks/subgraphClients/usePoolTogetherSubgraphClient'
