+++ lib/hooks/usePoolTokenData.js
@@ -1,13 +1,14 @@
-import { batch, contract } from '@pooltogether/etherplex'
-import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
-import DelegateableERC20ABI from 'abis/DelegateableERC20ABI'
-import { CONTRACT_ADDRESSES, QUERY_KEYS } from 'lib/constants'
-import { useContext } from 'react'
+import { useContext, useEffect } from 'react'
 import { useQuery } from 'react-query'
 import { ethers } from 'ethers'
-import { useReadProvider } from 'lib/hooks/useReadProvider'
 import { atom, useAtom } from 'jotai'
-import { useEffect } from 'react'
+import { batch, contract } from '@pooltogether/etherplex'
+
+import DelegateableERC20ABI from 'abis/DelegateableERC20ABI'
+import { CONTRACT_ADDRESSES } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
+import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
+import { useReadProvider } from 'lib/hooks/useReadProvider'
 
 const poolTokenDataAtom = atom({})
