+++ lib/hooks/useRetroactivePoolClaimData.js
@@ -5,7 +5,8 @@ import { ethers } from 'ethers'
 
 import MerkleDistributorAbi from 'abis/MerkleDistributor'
 
-import { CONTRACT_ADDRESSES, QUERY_KEYS } from 'lib/constants'
+import { CONTRACT_ADDRESSES } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { axiosInstance } from 'lib/axiosInstance'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { useReadProvider } from 'lib/hooks/useReadProvider'
