+++ lib/hooks/useTokenFaucetDripRate.js
@@ -3,7 +3,7 @@ import { batch, contract } from '@pooltogether/etherplex'
 import { useQuery } from 'react-query'
 import { useContext } from 'react'
 
-import { QUERY_KEYS } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { useReadProvider } from 'lib/hooks/useReadProvider'
 import { testAddress } from 'lib/utils/testAddress'
@@ -16,11 +16,7 @@ export const useTokenFaucetDripRate = (tokenFaucetAddress) => {
   const isZeroAddress = tokenFaucetAddress === '0x0000000000000000000000000000000000000000'
 
   const enabled =
-    tokenFaucetAddress &&
-    !pauseQueries &&
-    !addressError &&
-    !isZeroAddress &&
-    readProviderIsLoaded
+    tokenFaucetAddress && !pauseQueries && !addressError && !isZeroAddress && readProviderIsLoaded
 
   return useQuery(
     [QUERY_KEYS.tokenFaucetDripRateQuery, tokenFaucetAddress, chainId],
@@ -34,16 +30,12 @@ export const useTokenFaucetDripRate = (tokenFaucetAddress) => {
   )
 }
 
-async function getTokenFaucetDripRate (provider, tokenFaucetAddress) {
+async function getTokenFaucetDripRate(provider, tokenFaucetAddress) {
   const values = {}
 
   try {
     const tokenFaucetContract = contract('tokenFaucet', TokenFaucetABI, tokenFaucetAddress)
-    const { tokenFaucet } = await batch(
-      provider,
-      tokenFaucetContract
-        .dripRatePerSecond()
-    )
+    const { tokenFaucet } = await batch(provider, tokenFaucetContract.dripRatePerSecond())
 
     values.dripRatePerSecond = tokenFaucet.dripRatePerSecond[0]
   } catch (e) {
