+++ lib/hooks/useTotalPoolPrizeInterestUSD.js
@@ -1,4 +1,4 @@
-import { QUERY_KEYS } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { useQuery } from 'react-query'
 import { batch, contract } from '@pooltogether/etherplex'
 import { isEmpty } from 'lodash'
