+++ lib/hooks/useUsersV2Query.js
@@ -2,7 +2,8 @@ import { useContext } from 'react'
 import { useQuery } from 'react-query'
 import { isEmpty } from 'lodash'
 
-import { MAINNET_POLLING_INTERVAL, QUERY_KEYS } from 'lib/constants'
+import { MAINNET_POLLING_INTERVAL } from 'lib/constants'
+import { QUERY_KEYS } from 'lib/constants/queryKeys'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { fetchUsersV2Data } from 'lib/utils/fetchUsersV2Data'
